dh037

+++ src/main/java/ch/ethz/idsc/amodeus/dispatcher/parking/RestrictedLinkCapacityDispatcher.java
@@ -77,7 +77,7 @@ public class RestrictedLinkCapacityDispatcher extends SharedRebalancingDispatche
         dispatchPeriod = safeConfig.getInteger("dispatchPeriod", 60);
         sharingPeriod = safeConfig.getInteger("sharingPeriod", 10); // makes sense to choose this value similar to the
                                                                     // pickup duration
-        freeParkingPeriod = (long) safeConfig.getInteger("freeParkingPeriod", 10);
+        freeParkingPeriod = safeConfig.getInteger("freeParkingPeriod", 10);
 
         double rMax = safeConfig.getDouble("rMax", 1000.0);
         double phiMax = Pi.in(1000).multiply(RealScalar.of(safeConfig.getDouble("phiMaxDeg", 5.0) / 180.0)).number().doubleValue();
